Add Jest tests and workflow
