+++ package.json
@@ -4,10 +4,13 @@
   "description": "Chrome Extension",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "dependencies": {
     "spellchecker": "^3.7.0"
+  },
+  "devDependencies": {
+    "jest": "^29.6.1"
   },
   "keywords": [],
   "author": "",
