+++ popup.html
@@ -41,6 +41,7 @@
       <label><input type="checkbox" id="spellToggle" checked> Enable Spell Correction</label>
     </div>
   </div>
+  <script src="./js/ocrUtils.js"></script>
   <script src="./js/popup.js"></script>
 </body>
 </html>
